Import “converted” objects as media attachments if possible
This allows Image and Audio objects at least to be imported if any of
the URLs holds a supported media type. Only the first such link is
imported, even if multiple URLs are associated with the object, since
they are generally alternative representations (e.g. different formats)
of the same object.